# Office-management-railway-deployment/swagger.yaml
openapi: 3.0.0
info:
  title: Office Management API
  version: 1.0.0
  description: API documentation for the Office Management System Backend
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://office-management-sbuh.onrender.com
    description: Production server (Render)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in a cookie
security:
  - cookieAuth: []
paths:
...
  /api/admins/addAdmin:
    post:
      summary: Add a new administrator
      tags:
        - Admin-Management
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: Admin created successfully
        "400":
          description: Validation error
        "409":
          description: Email already exists
        "500":
          description: Server error
  /api/admins/getAdmins:
    get:
      summary: Get all administrators
      tags:
        - Admin-Management
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of admins
        "404":
          description: No admins found
        "500":
          description: Server error
  /api/admins/removeAdmin/{id}:
    delete:
      summary: Remove an administrator
      tags:
        - Admin-Management
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admin removed successfully
        "404":
          description: Admin not found
        "500":
          description: Server error
  /api/admins/updateAdmin/{id}:
    put:
      summary: Update an administrator
      tags:
        - Admin-Management
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Admin updated successfully
        "404":
          description: Admin not found
        "500":
          description: Server error
...
  /auth/login:
    post:
      summary: User login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                forceLogout:
                  type: boolean
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials
        "500":
          description: Server error
...
  /auth/me:
    get:
      summary: Get currently logged in user info (from token)
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Decoded user info
        "401":
          description: Not authenticated or invalid token
  /auth/myData:
    get:
      summary: Get current user full profile from DB
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Full user profile
        "401":
          description: Not authenticated or invalid token
  /auth/logout:
    post:
      summary: Logout user and clear session
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logout successful
        "500":
          description: Server error
  /auth/2fa:
    get:
      summary: Check if 2FA is enabled for the current user
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 2FA status
  /auth/authorized-2fa:
    post:
      summary: Verify 2FA code during login
      tags:
        - Auth
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
      responses:
        "200":
          description: 2FA verification successful
        "401":
          description: Invalid 2FA code or unauthorized
        "500":
          description: Server error
  /auth/verify-password:
    post:
      summary: Verify current user password
      tags:
        - Auth
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Password verified
        "401":
          description: Invalid password or unauthorized
        "500":
          description: Server error
...